# docs/schema.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "schema/v1/constant.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/v1/login": {
      "post": {
        "operationId": "AuthSchema_Login",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1AuthResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1LoginRequest"
            }
          }
        ],
        "tags": [
          "v1.AuthSchema"
        ]
      }
    },
...
    "/api/v1/register": {
      "post": {
        "operationId": "AuthSchema_Register",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1AuthResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1RegisterRequest"
            }
          }
        ],
        "tags": [
          "v1.AuthSchema"
        ]
      }
    },
...
  },
  "definitions": {
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1AuthResponse": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/v1AuthResponseData"
        }
      }
    },
    "v1AuthResponseData": {
      "type": "object",
      "properties": {
        "access_token": {
          "type": "string"
        }
      }
    },
...
    "v1Gender": {
      "type": "string",
      "enum": [
        "UNKNOWN_GENDER",
        "MALE",
        "FEMALE"
      ],
      "default": "UNKNOWN_GENDER"
    },
...
    "v1LoginRequest": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
...
    "v1RegisterRequest": {
      "type": "object",
      "properties": {
        "full_name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "confirm_password": {
          "type": "string"
        },
        "birth_date": {
          "type": "string"
        },
        "gender": {
          "$ref": "#/definitions/v1Gender"
        },
        "company": {
          "type": "string"
        },
        "job_title": {
          "type": "string"
        }
      }
    },
...
  }
}